[modify]查無資料時，不要顯示"轉檔完成!"

+++ PhoenixCI/FormUI/Prefix4/W40050.cs
@@ -29,14 +29,8 @@ namespace PhoenixCI.FormUI.Prefix4
 
          OCFG daoOCFG = new OCFG();
          oswGrpLookItem.SetDataTable(daoOCFG.ListAll(), "OSW_GRP", "OSW_GRP_NAME", DevExpress.XtraEditors.Controls.TextEditStyles.DisableTextEditor, null);
+         //取得交易時段決定下拉選單的值
          oswGrpLookItem.EditValue = daoOCFG.f_gen_osw_grp();
-      }
-
-      public override ResultStatus BeforeOpen()
-      {
-         base.BeforeOpen();
-
-         return ResultStatus.Success;
       }
 
       protected override ResultStatus Open()
@@ -95,19 +89,13 @@ namespace PhoenixCI.FormUI.Prefix4
          Thread.Sleep(5);
       }
 
-      private string OutputShowMessage {
-         set {
-            if (value != MessageDisplay.MSG_OK)
-               MessageDisplay.Info(value);
-         }
-      }
-
       protected override ResultStatus Export()
       {
          if (!StartExport()) {
             return ResultStatus.Fail;
          }
          string saveFilePath = string.Empty;
+         
          try {
             //資料來源
             DataTable dt = new D40050().GetData(emDate.Text.AsDateTime(), (emYear.Text + "/01/01").AsDateTime(), $"{oswGrpLookItem.EditValue.AsString()}%");
@@ -115,24 +103,34 @@ namespace PhoenixCI.FormUI.Prefix4
                MessageDisplay.Info(emDate.Text + ",讀取「保證金歷次調整紀錄」無任何資料!");
                return ResultStatus.Success;
             }
+            //複製template
             saveFilePath = PbFunc.wf_copy_file(_ProgramID, "40050");
+            //呼叫邏輯類別
             b40050 = new B40050(saveFilePath, dt, emCount.Text.AsInt());
-
+            MessageDisplay message = new MessageDisplay();
             /* Sheet:期貨data */
             ShowMsg("40051－期貨資料 轉檔中...");
-            OutputShowMessage = b40050.Wf40051();
+            message.OutputShowMessage = b40050.Wf40051();
             /* Sheet:選擇權data */
             ShowMsg("40052－選擇權資料 轉檔中...");
-            OutputShowMessage = b40050.Wf40052();
+            message.OutputShowMessage = b40050.Wf40052();
             /* Sheet:期貨股票類data */
             ShowMsg("40053－期貨股票類資料 轉檔中...");
-            OutputShowMessage = b40050.Wf40053();
+            message.OutputShowMessage = b40050.Wf40053();
             /* Sheet:選擇權股票類data */
             ShowMsg("40054－選擇權股票類資料 轉檔中...");
-            OutputShowMessage = b40050.Wf40054();
+            message.OutputShowMessage = b40050.Wf40054();
+
+            //全部查無資料刪除檔案
+            if (string.IsNullOrEmpty(message.OutputShowMessage)) {
+               if (File.Exists(saveFilePath))
+                  File.Delete(saveFilePath);
+               return ResultStatus.Fail;
+            }
          }
          catch (Exception ex) {
-            File.Delete(saveFilePath);
+            if (File.Exists(saveFilePath))
+               File.Delete(saveFilePath);
             WriteLog(ex);
             return ResultStatus.Fail;
          }
