+++ PhoenixCI/FormUI/Prefix4/W40060.cs
@@ -28,14 +28,8 @@ namespace PhoenixCI.FormUI.Prefix4
          this.Text = _ProgramID + "─" + _ProgramName;
          OCFG daoOCFG = new OCFG();
          oswGrpLookItem.SetDataTable(daoOCFG.ListAll(), "OSW_GRP", "OSW_GRP_NAME", DevExpress.XtraEditors.Controls.TextEditStyles.DisableTextEditor, null);
+         //取得交易時段決定下拉選單的值
          oswGrpLookItem.EditValue = daoOCFG.f_gen_osw_grp();
-      }
-
-      public override ResultStatus BeforeOpen()
-      {
-         base.BeforeOpen();
-
-         return ResultStatus.Success;
       }
 
       protected override ResultStatus Open()
@@ -94,13 +88,6 @@ namespace PhoenixCI.FormUI.Prefix4
          Thread.Sleep(5);
       }
 
-      private string OutputShowMessage {
-         set {
-            if (value != MessageDisplay.MSG_OK)
-               MessageDisplay.Info(value);
-         }
-      }
-
       protected override ResultStatus Export()
       {
          if (!StartExport()) {
@@ -114,21 +101,31 @@ namespace PhoenixCI.FormUI.Prefix4
                MessageDisplay.Info(emDate.Text + ",讀取「Span歷次調整紀錄」無任何資料!");
                return ResultStatus.Success;
             }
+            //複製template
             saveFilePath = PbFunc.wf_copy_file(_ProgramID, "40060");
+            //呼叫邏輯類別
             b40060 = new B40060(saveFilePath, dt, emCount.Text.AsInt());
-
+            MessageDisplay message = new MessageDisplay();
             /* Sheet:VSR */
             ShowMsg("40061－VSR 轉檔中...");
-            OutputShowMessage = b40060.Wf40061();
+            message.OutputShowMessage = b40060.Wf40061();
             /* Sheet:Spread Credit */
             ShowMsg("40062－Spread Credit 轉檔中...");
-            OutputShowMessage = b40060.Wf40062();
+            message.OutputShowMessage = b40060.Wf40062();
             /* Sheet:Delta Per Spread Ratio */
             ShowMsg("40063－Delta Per Spread Ratio 轉檔中...");
-            OutputShowMessage = b40060.Wf40063();
+            message.OutputShowMessage = b40060.Wf40063();
+
+            //全部查無資料刪除檔案
+            if (string.IsNullOrEmpty(message.OutputShowMessage)) {
+               if (File.Exists(saveFilePath))
+                  File.Delete(saveFilePath);
+               return ResultStatus.Fail;
+            }
          }
          catch (Exception ex) {
-            File.Delete(saveFilePath);
+            if (File.Exists(saveFilePath))
+               File.Delete(saveFilePath);
             WriteLog(ex);
             return ResultStatus.Fail;
          }
